+++ config/index.js
@@ -1,0 +1,9 @@
+
+let config;
+if (process.env.NODE_ENV === 'test') {
+	config = require('./config.unittest.js').default;
+} else {
+	config = require('./config.js').default;
+}
+
+export default config;