+++ src/cmp.stub.js
@@ -1,58 +1,66 @@
-import listener from "./cmp.ssp";
-import Config from "./lib/config";
-import log from './lib/log';
-import {hasLocalVendorConsentCookie} from './lib/cookie/cookie';
-import {GeolocationClient, GDPR_APPLIES_NO} from "./lib/geolocation/client";
-import {GeolocationEventManager} from "./lib/geolocation/eventManager";
+import listener from "./lib/ssp-fallback";
+import config from '../config';
 
-(function(window) {
-	window.ndmCmpConfig = window.ndmCmpConfig || {};
-	const ndmCmpConfig = window.ndmCmpConfig;
-	const configUpdates = {
-		...ndmCmpConfig
-	};
-	Config.update(configUpdates);
+(function() {
+	const element = document.createElement('script');
+	const firstScript = document.getElementsByTagName('script')[0];
+	const milliseconds = (new Date).getTime();
 
-	window.ndmCmpGeolocationEventManager = new GeolocationEventManager();
-
-	if (!hasLocalVendorConsentCookie()) {
-		const geolocationClient = new GeolocationClient(Config.geolocation);
-		geolocationClient.gdprCheck()
-			.then(result => {
-				log.debug(`GDPR check result '${result}'`);
-				window.ndmCmpConfig.gdprApplies = result !== GDPR_APPLIES_NO;
-				window.ndmCmpGeolocationEventManager.gdprCheckDone();
-			});
+	const defaultRequestUrl = 'https://quantcast.mgr.consensu.org/choice.js';
+	if (config.quantcast && config.quantcast.choiceID) {
+		const xhttp = new XMLHttpRequest();
+		const host = window.location.hostname;
+		const url = 'https://quantcast.mgr.consensu.org'
+			.concat('/choice/', config.quantcast.choiceID, '/', host, '/choice.js')
+			.concat('?timestamp=', milliseconds);
+		xhttp.onreadystatechange = function() {
+			if (this.readyState === 4) {
+				element.async = true;
+				element.type = 'text/javascript';
+				if (this.status === 200) {
+					element.src = url;
+				} else {
+					element.src = defaultRequestUrl;
+				}
+				firstScript.parentNode.insertBefore(element, firstScript);
+			}
+		};
+		xhttp.open('GET', url, true);
+		xhttp.send();
 	} else {
-		window.ndmCmpGeolocationEventManager.gdprCheckDone();
+		console.warn('QuantCast choice ID is not set. Using default choice script.');
+		element.src = defaultRequestUrl;
+		firstScript.parentNode.insertBefore(element, firstScript);
 	}
-
-	const commandQueue = [];
-	const cmp = function (command, parameter, callback) {
-		commandQueue.push({
-			command,
-			parameter,
-			callback
-		});
-	};
-	cmp.commandQueue = commandQueue;
-	cmp.receiveMessage = function (event) {
-		const data = event && event.data && event.data.__cmpCall;
-		if (data) {
-			const {callId, command, parameter} = data;
-			commandQueue.push({
-				callId,
-				command,
-				parameter,
-				event
-			});
+})();
+if (typeof window.__cmp === 'undefined') {
+	let count = 0;
+	window.__cmp = function() {
+		const arg = arguments;
+		if (typeof window.__cmp.a !== 'object') {
+			if (count < 10) {
+				setTimeout(() => {
+					window.__cmp.apply(window.__cmp, arg);
+				}, 400);
+				count ++;
+			} else {
+				console.warn('CMP not loaded after 4 seconds');
+			}
+		} else {
+			return window.__cmp.apply(window.__cmp, arg);
 		}
 	};
+}
 
-	window.__cmp = cmp;
+window.ndmtag = window.ndmtag || {};
+window.ndmtag.cmd = window.ndmtag.cmd || [];
 
-	window.ndmtag = window.ndmtag || {};
-	window.ndmtag.cmd = window.ndmtag.cmd || [];
+window.addEventListener("message", listener, false);
 
-	window.addEventListener("message", listener, false);
-}(window));
+// fetch google personalization consent information
+window.__cmp('getGooglePersonalization', (consent, isSuccess) => {
+	// request non personalized ads if we don't have a cookie or if no consent is given
+	if (!isSuccess || !consent.googlePersonalizationData.consentValue) {
+		(window.adsbygoogle=window.adsbygoogle||[]).requestNonPersonalizedAds=1;
+	}
+});