+++ src/lib/log.test.js
@@ -1,6 +1,4 @@
 import { expect } from 'chai';
-
-import config from './config';
 import log from './log';
 
 describe('log', () => {
@@ -12,8 +10,38 @@ describe('log', () => {
 		window.console.error = jest.fn();
 	});
 
-	it('defaults to not logging', () => {
+	it('default to warn', () => {
+		expect(log.logLevel).to.equal('warn');
+	});
 
+	it('use console.log for "debug"', () => {
+		log.logLevel = 'debug';
+		log.debug('debug');
+		log.info('info');
+		log.warn('warn');
+		log.error('error');
+
+		expect(window.console.log.mock.calls).to.have.length(1);
+		expect(window.console.info.mock.calls).to.have.length(1);
+		expect(window.console.warn.mock.calls).to.have.length(1);
+		expect(window.console.error.mock.calls).to.have.length(1);
+	});
+
+	it('use console.info for "info"', () => {
+		log.logLevel = 'info';
+		log.debug('debug');
+		log.info('info');
+		log.warn('warn');
+		log.error('error');
+
+		expect(window.console.log.mock.calls).to.be.empty;
+		expect(window.console.info.mock.calls).to.have.length(1);
+		expect(window.console.warn.mock.calls).to.have.length(1);
+		expect(window.console.error.mock.calls).to.have.length(1);
+	});
+
+	it('use console.warn for "warn"', () => {
+		log.logLevel = 'warn';
 		log.debug('debug');
 		log.info('info');
 		log.warn('warn');
@@ -21,25 +49,12 @@ describe('log', () => {
 
 		expect(window.console.log.mock.calls).to.be.empty;
 		expect(window.console.info.mock.calls).to.be.empty;
-		expect(window.console.warn.mock.calls).to.be.empty;
-		expect(window.console.error.mock.calls).to.be.empty;
+		expect(window.console.warn.mock.calls).to.have.length(1);
+		expect(window.console.error.mock.calls).to.have.length(1);
 	});
 
-	it('use console.log for "debug"', () => {
-		config.update({
-			logging: 'debug'
-		});
-
-		log.debug('debug');
-
-		expect(window.console.log.mock.calls).to.have.length(1);
-	});
-
-	it('respects logging level "error"', () => {
-		config.update({
-			logging: 'error'
-		});
-
+	it('use console.error for "error"', () => {
+		log.logLevel = 'error';
 		log.debug('debug');
 		log.info('info');
 		log.warn('warn');